+++ Datamigratie.Server/Features/MigrateZaken/ManageMigrations/GetMigration/Models/MigrationStatusResponse.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion.Internal;
 
-namespace Datamigratie.Server.Features.Migrate.ManageMigrations.StartMigration.Models
+namespace Datamigratie.Server.Features.MigrateZaken.ManageMigrations.GetMigration.Models
 {
     public class MigrationStatusResponse
     {