+++ Datamigratie.Server/Features/MigrateZaken/ManageMigrations/ShowMigrationHistory/GetZakenMigrationHistory/GetMigrationRecordsController.cs
@@ -1,8 +1,8 @@
-﻿using Datamigratie.Server.Features.Migrate.ManageMigrations.MigrationHistory.GetZakenMigrationHistory.Models;
-using Datamigratie.Server.Features.Migrate.ManageMigrations.MigrationHistory.GetZakenMigrationHistory.Services;
+﻿using Datamigratie.Server.Features.MigrateZaken.ManageMigrations.ShowMigrationHistory.GetZakenMigrationHistory.Models;
+using Datamigratie.Server.Features.MigrateZaken.ManageMigrations.ShowMigrationHistory.GetZakenMigrationHistory.Services;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Datamigratie.Server.Features.Migrate.ManageMigrations.MigrationHistory.GetZakenMigrationHistory;
+namespace Datamigratie.Server.Features.MigrateZaken.ManageMigrations.ShowMigrationHistory.GetZakenMigrationHistory;
 
 [ApiController]
 [Route("api/migration")]